requirements.txt
streamlit==1.26.0
google-generativeai==0.3.1
